requirements.txt
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
et-xmlfile==1.1.0
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
numpy==1.24.1
openai==0.25.0
openapi==1.1.0
openpyxl==3.0.10
pandas==1.5.2
pandas-stubs==1.5.2.221213
pyrsistent==0.19.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.1
six==1.16.0
tqdm==4.64.1
types-pytz==2022.7.0.0
typing_extensions==4.4.0
urllib3==1.26.13
